==> NLP/dynamic_topic_modeling_of_tech_news/05_Embeddings.ipynb <==
# Create Embeddings
To speed up the D-ETM training process, the construction of the word embedding matrices is placed in this separate notebook. It includes the creation of (smaller) embedding matrices for Word2Vec, GloVe and fastText. They include embeddings for bigrams and trigrams.

## Directories & Libraries

from google.colab import drive
drive.mount('/content/drive', force_remount= True)

%cd '/content/drive/My Drive/Thesis/Topic-Modeling/'
data_path = 'Data/Technology-Data/processed/preprocessed/texts.txt'

!pip install fasttext

import fasttext
import fasttext.util
import gensim
import numpy as np
import os
import time
import utils
import os

## Settings

emb_size = 300

## Get Vocabulary

vocab, full, train, valid, test = utils.get_data('Data/Technology-Data/processed/final/grouped_years/min_df_50')

len(vocab)

## Word2Vec

docs = []
with open(data_path) as articles:
    a = articles.readlines()
    for line in a:
        tokens = line.split()
        docs.append(tokens)

start_time = time.time()
model_word2vec = gensim.models.Word2Vec(docs,
                                        min_count=50,
                                        sg=1,
                                        size=emb_size,
                                        iter=100,
                                        workers=6,
                                        negative=10,
                                        window=5)

calc_time = time.time() - start_time
print("--- %s seconds ---" % calc_time)

word2vec_file = 'Data/Embeddings/Word2Vec/Word2Vec_{}'.format(emb_size)

model_word2vec.save(word2vec_file + '.model')

with open(word2vec_file + '.txt','w') as f:
    for word in list(model_word2vec.wv.vocab):
        vector = list(model_word2vec.wv.__getitem__(word))
        vector_str = " ".join(['%.9f' % val for val in vector])
        f.write(word + ' ' + vector_str + '\n')

## GloVe

Load original cased (!) GloVe vectors (https://nlp.stanford.edu/projects/glove/, 11/2020)

glove_orig = {}
with open('Data/Embeddings/GloVe/glove.840B.300d.txt', 'rb') as f:
    for l in f:
        line = l.decode().split()
        word = line[0]
        try:
            vect = np.array(line[1:]).astype(np.float)
        except ValueError:
            if len(line[1:])==301: # a few GloVe lines contain more than 1 string token
                vect = np.array(line[2:]).astype(np.float)
        glove_orig[word] = vect

len(glove_orig)

Create data-/task-specific GloVe embeddings by iterating through the vocabulary. Account for bigrams and trigrams by averaging word vectors

glove_embeddings = {}
words_no_glove_emb = []
for v in vocab:
    try:
        tokens = v.split('_')
        if len(tokens) == 3:
            glove_embeddings[v] = np.mean(np.array([glove_orig[tokens[0]],glove_orig[tokens[1]],glove_orig[tokens[2]]]), axis=0)
        elif len(tokens) == 2:
            glove_embeddings[v] = np.mean(np.array([glove_orig[tokens[0]],glove_orig[tokens[1]]]), axis=0)
        else:
            glove_embeddings[v] = glove_orig[v]
    except KeyError:
        words_no_glove_emb.append(v)
        
print('No embedding available for:', ', '.join(words_no_glove_emb))

Save data/task-specific GloVe word embeddings

with open('Data/Embeddings/GloVe/GloVe_300.txt', 'w') as f:
  for word in glove_embeddings.keys():
      vec = ['%.9f' % val for val in glove_embeddings[word]]
      vec_str = " ".join(vec)
      emb_line = word + " " + vec_str + '\n'
      f.write(emb_line)

## fastText
https://fasttext.cc/, https://fasttext.cc/docs/en/crawl-vectors.html (11/2020)

#fasttext.util.download_model('en', if_exists='ignore') # download the model if it has not been done before

fastText_model = fasttext.load_model('Data/Embeddings/fastText/cc.en.300.bin')
fastText_model.get_dimension()
if emb_size < 300:
    fasttext.util.reduce_model(fastText_model, emb_size)
fastText_model.get_dimension()    

print('Neighbouring Example for ai: ', fastText_model.get_nearest_neighbors("ai"))
print('Neighbouring Example for AI: ', fastText_model.get_nearest_neighbors("AI"))

print('Neighbouring Example for Apple: ', fastText_model.get_nearest_neighbors("Apple"))
print('Neighbouring Example for apple: ', fastText_model.get_nearest_neighbors("apple"))

print('Analogy Example for PS1, PS2, 4G: ', fastText_model.get_analogies("PS1", "PS2", "4G"))

Create a subset of data-/task-specific fastText embeddings by iterating through the vocabulary. Account for bigrams and trigrams by applying fastText's get_sentence_vector():

words_no_ft_emb = []
fasttext_embeddings = {}

for word in vocab:
    try:  
      # Bigrams, Trigrams:
        if '_' in word:
            fasttext_embeddings[word] = fastText_model.get_sentence_vector(word.replace('_',' '))
        else:
            fasttext_embeddings[word] = fastText_model.get_word_vector(word)

    except Exception as e:
        words_no_ft_emb.append(word)

print('Number of words for which no fastText embedding could not be received:', len(words_no_ft_emb))
if len(words_no_ft_emb)<15:
    print('Corresponding words:', words_no_ft_emb)
del fastText_model

fasttext_filename = 'Data/Embeddings/fastText/fastText_{}.txt'.format(emb_size)
with open(fasttext_filename, 'w') as f:
    for word in fasttext_embeddings.keys():
        vec = ['%.9f' % val for val in fasttext_embeddings[word]]
        vec_str = " ".join(vec)
        emb_line = word + " " + vec_str + '\n'
        f.write(emb_line)
